Cardano.Protocol.TPraos.API

Description

Integration between the Shelley ledger and its corresponding (Transitional Praos) protocol.

In particular, this code supports extracting the components of the ledger state needed for protocol execution, both now and in a 2k-slot window.

Documentation

class (Crypto c, DSignable c (OCertSignable c), KESignable c (BHBody c), VRFSignable c Seed) => PraosCrypto c

class (ChainData (ChainDepState (Crypto era)), SerialisableData (ChainDepState (Crypto era)), Eq (ChainTransitionError (Crypto era)), Show (ChainTransitionError (Crypto era)), Show (LedgerView (Crypto era)), Show (FutureLedgerViewError era), STS (EraRule "TICKF" era), BaseM (EraRule "TICKF" era) ~ ShelleyBase, Environment (EraRule "TICKF" era) ~ (), State (EraRule "TICKF" era) ~ NewEpochState era, Signal (EraRule "TICKF" era) ~ SlotNo, PredicateFailure (EraRule "TICKF" era) ~ TickfPredicateFailure era, HasField "_d" (PParams era) UnitInterval, HasField "_maxBBSize" (PParams era) Natural, HasField "_maxBHSize" (PParams era) Natural, HasField "_protocolVersion" (PParams era) ProtVer) => GetLedgerView era where

Minimal complete definition

Nothing

Methods

currentLedgerView :: NewEpochState era -> LedgerView (Crypto era)

default currentLedgerView :: HasField "_extraEntropy" (PParams era) Nonce => NewEpochState era -> LedgerView (Crypto era)

futureLedgerView :: MonadError (FutureLedgerViewError era) m => Globals -> NewEpochState era -> SlotNo -> m (LedgerView (Crypto era))

default futureLedgerView :: (MonadError (FutureLedgerViewError era) m, HasField "_extraEntropy" (PParams era) Nonce) => Globals -> NewEpochState era -> SlotNo -> m (LedgerView (Crypto era))

data LedgerView crypto

Data required by the Transitional Praos protocol from the Shelley ledger.

Constructors

LedgerView

Fields

lvD :: UnitInterval

lvExtraEntropy :: ~Nonce

lvPoolDistr :: PoolDistr crypto

lvGenDelegs :: GenDelegs crypto

lvChainChecks :: ChainChecksPParams

mkInitialShelleyLedgerView :: forall c. ShelleyGenesis (ShelleyEra c) -> LedgerView c

We construct a LedgerView using the Shelley genesis config in the same way as translateToShelleyLedgerState.

newtype FutureLedgerViewError era

Constructors

FutureLedgerViewError [PredicateFailure (EraRule "TICKF" era)]

Chain state operations

The chain state is an amalgam of the protocol state and the ticked nonce.

data ChainDepState crypto

Constructors

ChainDepState

Fields

csProtocol :: !(PrtclState crypto)

csTickn :: !TicknState

csLabNonce :: !Nonce

Nonce constructed from the hash of the last applied block header.

newtype ChainTransitionError crypto

Constructors

ChainTransitionError [PredicateFailure (PRTCL crypto)]

tickChainDepState

Arguments

:: Globals

-> LedgerView crypto

-> Bool
Are we in a new epoch?

-> ChainDepState crypto

-> ChainDepState crypto

Tick the chain state to a new epoch.

updateChainDepState :: forall crypto m. (PraosCrypto crypto, MonadError (ChainTransitionError crypto) m) => Globals -> LedgerView crypto -> BHeader crypto -> ChainDepState crypto -> m (ChainDepState crypto)

Update the chain state based upon a new block header.

This also updates the last applied block hash.

reupdateChainDepState :: forall crypto. PraosCrypto crypto => Globals -> LedgerView crypto -> BHeader crypto -> ChainDepState crypto -> ChainDepState crypto

Re-update the chain state based upon a new block header.

This function does no validation of whether the header is internally valid or consistent with the chain it is being applied to; the caller must ensure that this is valid through having previously applied it.

initialChainDepState :: Nonce -> Map (KeyHash 'Genesis crypto) (GenDelegPair crypto) -> ChainDepState crypto

Construct an initial chain state given an initial nonce and a set of genesis delegates.

checkLeaderValue :: forall v. VRFAlgorithm v => OutputVRF v -> Rational -> ActiveSlotCoeff -> Bool

Check that the certified VRF output, when used as a natural, is valid for being slot leader.

getLeaderSchedule :: (Era era, Signable (VRF (Crypto era)) Seed, HasField "_d" (PParams era) UnitInterval) => Globals -> NewEpochState era -> ChainDepState (Crypto era) -> KeyHash 'StakePool (Crypto era) -> SignKeyVRF (Crypto era) -> PParams era -> Set SlotNo

Get the (private) leader schedule for this epoch.

Given a private VRF key, returns the set of slots in which this node is eligible to lead.