Cardano.Protocol.TPraos.Rules.Prtcl

Documentation

data PRTCL c

type family State a

Type of the state which the system transitions between.

data PrtclEnv

Constructors

PrtclEnv UnitInterval PoolDistr GenDelegs Nonce

data PrtclState

Constructors

PrtclState

Fields

!(Map (KeyHash 'BlockIssuer) Word64)

Operation Certificate counters

!Nonce

Evolving nonce

!Nonce

Candidate nonce

data PrtclPredicateFailure c

Constructors

OverlayFailure (PredicateFailure (OVERLAY c))

UpdnFailure (PredicateFailure (UPDN c))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data PrtlSeqFailure

Constructors

WrongSlotIntervalPrtclSeq

Fields

SlotNo

Last slot number.

SlotNo

Current slot number.

WrongBlockNoPrtclSeq

Fields

(WithOrigin LastAppliedBlock)

Last applied block.

BlockNo

Current block number.

WrongBlockSequencePrtclSeq

Fields

PrevHash

Last applied hash

PrevHash

Current block's previous hash

prtlSeqChecks ∷ (MonadError PrtlSeqFailure m, Crypto c) ⇒ WithOrigin LastAppliedBlock → BHeader c → m ()