Test.Cardano.Protocol.Crypto.VRF.Fake

Description

Fake implementation of VRF, where the random value isn't random but given by the creator.

Documentation

newtype NatNonce

We provide our own nonces to mkBlock, which we then wish to recover as the output of the VRF functions. In general, however, we just derive them from a natural. Since the nonce is a hash, we do not want to recover it to find a preimage. In testing, therefore, we just wrap the raw natural, which we then encode into the fake VRF implementation.

Constructors

NatNonce Natural

data FakeVRF

data family VerKeyVRF v

data family SignKeyVRF v

data WithResult a

Constructors

WithResult !a !Word64