Test.Cardano.Protocol.TPraos.Arbitrary

Documentation

genBHeader ∷ (Signable (DSIGN c) (OCertSignable c), Signable (VRF c) Seed, Signable (KES c) (BHBody c), Crypto c) ⇒ [AllIssuerKeys c r] → Gen (BHeader c)

genBlock ∷ (Signable (DSIGN c) (OCertSignable c), Signable (VRF c) Seed, Signable (KES c) (BHBody c), EraSegWits era, Arbitrary (Tx era), c ~ EraCrypto era) ⇒ [AllIssuerKeys c r] → Gen (Block (BHeader c) era)

Use supplied keys to generate a Block.

genCoherentBlock ∷ ∀ era r. (EraSegWits era, Arbitrary (Tx era), Signable (KES (EraCrypto era)) ~ SignableRepresentation, Signable (DSIGN (EraCrypto era)) ~ SignableRepresentation, PraosCrypto (EraCrypto era)) ⇒ [AllIssuerKeys (EraCrypto era) r] → Gen (Block (BHeader (EraCrypto era)) era)

For some purposes, a totally random block generator may not be suitable. There are tests in the ouroboros-network repository, for instance, that perform some integrity checks on the generated blocks.

For other purposes, such as the serialization tests in this repository, genBlock is more appropriate.

This generator uses mkBlock provide more coherent blocks.

Orphan instances

Arbitrary KESPeriod

Instance details

Methods

arbitrary ∷ Gen KESPeriod

shrink ∷ KESPeriod → [KESPeriod]

Arbitrary TicknState

Instance details

Methods

arbitrary ∷ Gen TicknState

shrink ∷ TicknState → [TicknState]

(Crypto c, Signable (VRF c) ~ SignableRepresentation) ⇒ Arbitrary (BHBody c)

Instance details

Methods

arbitrary ∷ Gen (BHBody c)

shrink ∷ BHBody c → [BHBody c]

(Crypto c, Signable (VRF c) ~ SignableRepresentation, Signable (KES c) ~ SignableRepresentation) ⇒ Arbitrary (BHeader c)

Instance details

Methods

arbitrary ∷ Gen (BHeader c)

shrink ∷ BHeader c → [BHeader c]

Crypto c ⇒ Arbitrary (HashHeader c)

Instance details

Methods

arbitrary ∷ Gen (HashHeader c)

shrink ∷ HashHeader c → [HashHeader c]

Crypto c ⇒ Arbitrary (PrevHash c)

Instance details

Methods

arbitrary ∷ Gen (PrevHash c)

shrink ∷ PrevHash c → [PrevHash c]

Crypto c ⇒ Arbitrary (OCert c)

Instance details

Methods

arbitrary ∷ Gen (OCert c)

shrink ∷ OCert c → [OCert c]

Crypto c ⇒ Arbitrary (OBftSlot c)

Instance details

Methods

arbitrary ∷ Gen (OBftSlot c)

shrink ∷ OBftSlot c → [OBftSlot c]

Crypto c ⇒ Arbitrary (PrtclState c)

Instance details

Methods

arbitrary ∷ Gen (PrtclState c)

shrink ∷ PrtclState c → [PrtclState c]

(Era era, c ~ EraCrypto era, EraSegWits era, Signable (KES c) ~ SignableRepresentation, Signable (VRF c) ~ SignableRepresentation, Arbitrary (Tx era)) ⇒ Arbitrary (Block (BHeader c) era)

Instance details

Methods

arbitrary ∷ Gen (Block (BHeader c) era)

shrink ∷ Block (BHeader c) era → [Block (BHeader c) era]