Cardano.Protocol.TPraos.BHeader

Documentation

newtype HashHeader c

The hash of a Block Header

Constructors

HashHeader

Fields

unHashHeader ∷ Hash c EraIndependentBlockHeader

data PrevHash c

The previous hash of a block

Constructors

GenesisHash

BlockHash !(HashHeader c)

data BHeader c where

Bundled Patterns

pattern BHeader ∷ Crypto c ⇒ BHBody c → SignedKES c (BHBody c) → BHeader c

data BHBody c

Constructors

BHBody

Fields

bheaderBlockNo ∷ !BlockNo

block number

bheaderSlotNo ∷ !SlotNo

block slot

bheaderPrev ∷ !(PrevHash c)

Hash of the previous block header

bheaderVk ∷ !(VKey 'BlockIssuer c)

verification key of block issuer

bheaderVrfVk ∷ !(VerKeyVRF c)

VRF verification key for block issuer

bheaderEta ∷ !(CertifiedVRF c Nonce)

block nonce

bheaderL ∷ !(CertifiedVRF c Natural)

leader election value

bsize ∷ !Word32

Size of the block body

bhash ∷ !(Hash c EraIndependentBlockBody)

Hash of block body

bheaderOCert ∷ !(OCert c)

operational certificate

bprotver ∷ !ProtVer

protocol version

data LastAppliedBlock c

Constructors

LastAppliedBlock

Fields

labBlockNo ∷ !BlockNo

labSlotNo ∷ !SlotNo

labHash ∷ !(HashHeader c)

data BoundedNatural

Natural value with some additional bound. It must always be the base that 'bvValue <= bvMaxValue'. The creator is responsible for checking this value.

assertBoundedNatural

Arguments

∷ Natural
Maximum bound

→ Natural
Value

→ BoundedNatural

Assert that a natural is bounded by a certain value. Throws an error when this is not the case.

lastAppliedHash ∷ WithOrigin (LastAppliedBlock c) → PrevHash c

issuerIDfromBHBody ∷ Crypto c ⇒ BHBody c → KeyHash 'BlockIssuer c

Retrieve the issuer id (the hash of the cold key) from the body of the block header. This corresponds to either a genesis/core node or a stake pool.

checkLeaderValue ∷ ∀ v. VRFAlgorithm v ⇒ OutputVRF v → Rational → ActiveSlotCoeff → Bool

Check that the certified VRF output, when used as a natural, is valid for being slot leader.

checkLeaderNatValue

Arguments

∷ BoundedNatural
Certified nat value

→ Rational
Stake proportion

→ ActiveSlotCoeff

→ Bool

Check that the certified input natural is valid for being slot leader. This means we check that

p < 1 - (1 - f)^σ

where p = certNat / certNatMax.

The calculation is done using the following optimization:

let q = 1 - p and c = ln(1 - f)

then p < 1 - (1 - f)^σ = 1 / (1 - p) < exp(-σ * c) = 1 / q < exp(-σ * c)

This can be efficiently be computed by taylorExpCmp which returns ABOVE in case the reference value `1 / (1 - p)` is above the exponential function at `-σ * c`, BELOW if it is below or MaxReached if it couldn't conclusively compute this within the given iteration bounds.

Note that 1 1 1 certNatMax --- = ----- = ---------------------------- = ---------------------- q 1 - p 1 - (certNat / certNatMax) (certNatMax - certNat)

bhHash ∷ Crypto c ⇒ BHeader c → HashHeader c

Hash a given block header

hashHeaderToNonce ∷ HashHeader c → Nonce

HashHeader to Nonce What is going on here? This is here because the surrounding code is parametrized in the hash algorithm used, but the nonce is hard-coded to Blake2b_256. We require the nonce to have the right length (the size of a Blake2b_256 hash), so if the hash size differs, we pad or remove bytes accordingly.

prevHashToNonce ∷ PrevHash c → Nonce

bHeaderSize ∷ ∀ c. BHeader c → Int

bhbody ∷ Crypto c ⇒ BHeader c → BHBody c

hBbsize ∷ BHBody c → Word32

seedEta ∷ Nonce

seedL ∷ Nonce

mkSeed

Arguments

∷ Nonce
Universal constant

→ SlotNo

→ Nonce
Epoch nonce

→ Seed

Construct a seed to use in the VRF computation.

bnonce ∷ BHBody c → Nonce

Retrieve the new nonce from the block header body.

makeHeaderView ∷ Crypto c ⇒ BHeader c → BHeaderView c