Cardano.Protocol.TPraos.Rules.OCert

Documentation

data OCERT c

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data OCertEnv

Constructors

OCertEnv

Fields

ocertEnvStPools ∷ Set (KeyHash 'StakePool)

ocertEnvGenDelegs ∷ Set (KeyHash 'GenesisDelegate)

data OcertPredicateFailure

Constructors

KESBeforeStartOCERT KESPeriod KESPeriod

KESAfterEndOCERT KESPeriod KESPeriod Word64

CounterTooSmallOCERT Word64 Word64

InvalidSignatureOCERT Word64 KESPeriod

InvalidKesSignatureOCERT Word Word Word String

NoCounterForKeyHashOCERT (KeyHash 'BlockIssuer)