Cardano.Protocol.TPraos.Rules.Overlay

Documentation

data OVERLAY c

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data OverlayEnv

Constructors

OverlayEnv UnitInterval PoolDistr GenDelegs Nonce

data OverlayPredicateFailure c

Constructors

VRFKeyUnknown (KeyHash 'StakePool)

VRFKeyWrongVRFKey (KeyHash 'StakePool) (VRFVerKeyHash 'StakePoolVRF) (VRFVerKeyHash 'BlockIssuerVRF)

VRFKeyBadNonce Nonce SlotNo Nonce (CertifiedVRF (VRF c) Nonce)

VRFKeyBadLeaderValue Nonce SlotNo Nonce (CertifiedVRF (VRF c) Nonce)

VRFLeaderValueTooBig (OutputVRF (VRF c)) Rational ActiveSlotCoeff

NotActiveSlotOVERLAY SlotNo

WrongGenesisColdKeyOVERLAY (KeyHash 'BlockIssuer) (KeyHash 'GenesisDelegate)

WrongGenesisVRFKeyOVERLAY (KeyHash 'BlockIssuer) (VRFVerKeyHash 'GenDelegVRF) (VRFVerKeyHash 'BlockIssuerVRF)

UnknownGenesisKeyOVERLAY (KeyHash 'Genesis)

OcertFailure (PredicateFailure (OCERT c))

data OBftSlot

Constructors

NonActiveSlot

ActiveSlot !(KeyHash 'Genesis)

classifyOverlaySlot ∷ SlotNo → Set (KeyHash 'Genesis) → UnitInterval → ActiveSlotCoeff → SlotNo → OBftSlot

lookupInOverlaySchedule ∷ SlotNo → Set (KeyHash 'Genesis) → UnitInterval → ActiveSlotCoeff → SlotNo → Maybe OBftSlot

overlaySlots ∷ SlotNo → UnitInterval → EpochSize → [SlotNo]

Return the list of overlaySlots for a given epoch. Note that this linear in the size of the epoch, and should probably only be used for testing. If something more performant is needed, we could probably use [start + floor(xd) | x <- [0 .. (spe -1)], floor(xd) < spe] but we would need to make sure that this is equivalent.