Cardano.Protocol.TPraos.Rules.Tickn

Documentation

data TICKN

data TicknEnv

Constructors

TicknEnv

Fields

ticknEnvExtraEntropy :: Nonce

ticknEnvCandidateNonce :: Nonce

ticknEnvHashHeaderNonce :: Nonce

Hash of the last header of the previous epoch as a nonce.

data TicknState

Constructors

TicknState

Fields

ticknStateEpochNonce :: !Nonce

ticknStatePrevHashNonce :: !Nonce

data TicknPredicateFailure

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.